Added a block feature to the snipe command

+++ cogs/fun.py
@@ -177,24 +177,24 @@ class Fun(commands.Cog):
                     name=f"Replied to {ref.author.display_name} ({ref.author.id}) who said:",
                     value=ref.content,
                 )
-                embed.set_footer(text=f"React with 🚮 to delete this message.")
             except discord.errors.NotFound:
                 embed.set_footer(
                     text="Replying to a message that doesn't exist anymore. React with 🚮 to delete this message."
                 )
-
+                
         if not embed.footer:
-            embed.set_footer(text="React with 🚮 to delete this message.")
-
+            embed.set_footer(text="You bozo, you're blocked from deleting it!" if (ref.author.id in config["snipeblock"]) else "React with 🚮 to delete this message.")
+     
         snipemsg = await ctx.reply(f"Sniped message by {message.author.mention}", embed=embed)
         self.snipe_message = None
 
         def check(reaction, user):
             return user == message.author and str(reaction.emoji) == "🚮" and reaction.message == snipemsg
-
-        await snipemsg.add_reaction("🚮")
-        await self.bot.wait_for("reaction_add", timeout=60.0, check=check)
-        await snipemsg.delete()
+        
+        if not ref.author.id in config["snipeblock"]:
+            await snipemsg.add_reaction("🚮")
+            await self.bot.wait_for("reaction_add", timeout=60.0, check=check)
+            await snipemsg.delete()
 
 
 async def setup(bot):
